+++ Sources/SwiftDiscord/DiscordEndpointConsumer.swift
@@ -26,6 +26,11 @@ import Foundation
     that if you make modifications to the client inside of a callback, you first dispatch back on the handleQueue.
 */
 public protocol DiscordEndpointConsumer : DiscordUserActor {
+    // MARK: Properties
+
+    /// The rate limiter for this consumer.
+    var rateLimiter: DiscordRateLimiter { get }
+
     // MARK: Channels
 
     /**
@@ -477,253 +482,9 @@ public protocol DiscordEndpointConsumer : DiscordUserActor {
 
 public extension DiscordEndpointConsumer {
     /// Default implementation
-    public func acceptInvite(_ invite: String, callback: ((DiscordInvite?) -> ())? = nil) {
-        DiscordEndpoint.acceptInvite(invite, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func addPinnedMessage(_ messageId: String, on channelId: String, callback: ((Bool) -> ())? = nil) {
-        DiscordEndpoint.addPinnedMessage(messageId, on: channelId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func addGuildMemberRole(_ roleId: String, to userId: String, on guildId: String,
-                                   callback: ((Bool) -> ())?) {
-        DiscordEndpoint.addGuildMemberRole(roleId, to: userId, on: guildId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func bulkDeleteMessages(_ messages: [String], on channelId: String, callback: ((Bool) -> ())? = nil) {
-        DiscordEndpoint.bulkDeleteMessages(messages, on: channelId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func createDM(with: String, callback: @escaping (DiscordDMChannel?) -> ()) {
-        DiscordEndpoint.createDM(with: with, user: user!.id, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func createInvite(for channelId: String, options: [DiscordEndpoint.Options.CreateInvite],
-                             callback: @escaping (DiscordInvite?) -> ()) {
-        DiscordEndpoint.createInvite(for: channelId, options: options, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func createGuildChannel(on guildId: String, options: [DiscordEndpoint.Options.GuildCreateChannel],
-                                   callback: ((DiscordGuildChannel?) -> ())? = nil) {
-        DiscordEndpoint.createGuildChannel(guildId, options: options, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func createGuildRole(on guildId: String, withOptions options: [DiscordEndpoint.Options.CreateRole] = [],
-                                callback: @escaping (DiscordRole?) -> ()) {
-        DiscordEndpoint.createGuildRole(on: guildId, withOptions: options, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func createWebhook(forChannel channelId: String, options: [DiscordEndpoint.Options.WebhookOption],
-                              callback: @escaping (DiscordWebhook?) -> () = {_ in }) {
-        DiscordEndpoint.createWebhook(forChannel: channelId, with: token, options: options, callback: callback)
-    }
-
-    /// Default implementation
-    public func deleteChannel(_ channelId: String, callback: ((Bool) -> ())? = nil) {
-        DiscordEndpoint.deleteChannel(channelId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func deleteChannelPermission(_ overwriteId: String, on channelId: String, callback: ((Bool) -> ())? = nil) {
-        DiscordEndpoint.deleteChannelPermission(overwriteId, on: channelId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func deleteGuild(_ guildId: String, callback: ((DiscordGuild?) -> ())? = nil) {
-        DiscordEndpoint.deleteGuild(guildId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func deleteInvite(_ invite: String, callback: ((DiscordInvite?) -> ())? = nil) {
-        DiscordEndpoint.deleteInvite(invite, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func deleteMessage(_ messageId: String, on channelId: String, callback: ((Bool) -> ())? = nil) {
-        DiscordEndpoint.deleteMessage(messageId, on: channelId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func deletePinnedMessage(_ messageId: String, on channelId: String, callback: ((Bool) -> ())? = nil) {
-        DiscordEndpoint.deletePinnedMessage(messageId, on: channelId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func deleteWebhook(_ webhookId: String, callback: ((Bool) -> ())? = nil) {
-        DiscordEndpoint.deleteWebhook(webhookId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func editMessage(_ messageId: String, on channelId: String, content: String,
-                            callback: ((DiscordMessage?) -> ())? = nil) {
-        DiscordEndpoint.editMessage(messageId, on: channelId, content: content, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func editChannelPermission(_ permissionOverwrite: DiscordPermissionOverwrite, on channelId: String,
-                                      callback: ((Bool) -> ())? = nil) {
-        DiscordEndpoint.editChannelPermission(permissionOverwrite, on: channelId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func getChannel(_ channelId: String, callback: @escaping (DiscordGuildChannel?) -> ()) {
-        DiscordEndpoint.getChannel(channelId, with: token, callback: callback)
-    }
-
-    /// Default implementation
     public func getBotURL(with permissions: [DiscordPermission]) -> URL? {
         guard let user = self.user else { return nil }
 
         return DiscordOAuthEndpoint.createBotAddURL(for: user, with: permissions)
-    }
-
-    /// Default implementation
-    public func getDMs(callback: @escaping ([String: DiscordDMChannel]) -> ()) {
-        DiscordEndpoint.getDMs(user: user!.id, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func getGuildBans(for guildId: String, callback: @escaping ([DiscordBan]) -> ()) {
-        DiscordEndpoint.getGuildBans(for: guildId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func getGuildChannels(_ guildId: String, callback: @escaping ([DiscordGuildChannel]) -> ()) {
-        DiscordEndpoint.getGuildChannels(guildId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func getGuildMember(by id: String, on guildId: String, callback: @escaping (DiscordGuildMember?) -> ()) {
-        DiscordEndpoint.getGuildMember(by: id, on: guildId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func getGuildMembers(on guildId: String, options: [DiscordEndpoint.Options.GuildGetMembers],
-                                callback: @escaping ([DiscordGuildMember]) -> ()) {
-        DiscordEndpoint.getGuildMembers(on: guildId, options: options, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func getGuilds(callback: @escaping ([String: DiscordUserGuild]) -> ()) {
-        DiscordEndpoint.getGuilds(user: user!.id, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func getGuildRoles(for guildId: String, callback: @escaping ([DiscordRole]) -> ()) {
-        DiscordEndpoint.getGuildRoles(for: guildId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func getInvite(_ invite: String, callback: @escaping (DiscordInvite?) -> ()) {
-        DiscordEndpoint.getInvite(invite, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func getInvites(for channelId: String, callback: @escaping ([DiscordInvite]) -> ()) {
-        return DiscordEndpoint.getInvites(for: channelId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func getMessages(for channelId: String, options: [DiscordEndpoint.Options.GetMessage] = [],
-                            callback: @escaping ([DiscordMessage]) -> ()) {
-        DiscordEndpoint.getMessages(for: channelId, with: token, options: options, callback: callback)
-    }
-
-    /// Default implementation
-    public func getWebhook(_ webhookId: String, callback: @escaping (DiscordWebhook?) -> ()) {
-        DiscordEndpoint.getWebhook(webhookId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func getWebhooks(forChannel channelId: String, callback: @escaping ([DiscordWebhook]) -> ()) {
-        DiscordEndpoint.getWebhooks(forChannel: channelId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func getWebhooks(forGuild guildId: String, callback: @escaping ([DiscordWebhook]) -> ()) {
-        DiscordEndpoint.getWebhooks(forGuild: guildId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func guildBan(userId: String, on guildId: String, deleteMessageDays: Int = 7,
-                         callback: ((Bool) -> ())? = nil) {
-        DiscordEndpoint.guildBan(userId: userId, on: guildId, deleteMessageDays: deleteMessageDays,
-            with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func getPinnedMessages(for channelId: String, callback: @escaping ([DiscordMessage]) -> ()) {
-        DiscordEndpoint.getPinnedMessages(for: channelId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func modifyChannel(_ channelId: String, options: [DiscordEndpoint.Options.ModifyChannel],
-                              callback: ((DiscordGuildChannel?) -> ())? = nil) {
-        DiscordEndpoint.modifyChannel(channelId, options: options, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func modifyGuild(_ guildId: String, options: [DiscordEndpoint.Options.ModifyGuild],
-                            callback: ((DiscordGuild?) -> ())? = nil) {
-        DiscordEndpoint.modifyGuild(guildId, options: options, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func modifyGuildChannelPositions(on guildId: String, channelPositions: [[String: Any]],
-                                            callback: (([DiscordGuildChannel]) -> ())? = nil) {
-        DiscordEndpoint.modifyGuildChannelPositions(on: guildId, channelPositions: channelPositions,
-            with: token, callback: callback)
-    }
-
-    /// Default implementation
-    func modifyGuildMember(_ id: String, on guildId: String, options: [DiscordEndpoint.Options.ModifyMember],
-                           callback: ((Bool) -> ())? = nil) {
-        DiscordEndpoint.modifyGuildMember(id, on: guildId, options: options, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func modifyGuildRole(_ role: DiscordRole, on guildId: String, callback: ((DiscordRole?) -> ())? = nil) {
-        DiscordEndpoint.modifyGuildRole(role, on: guildId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func modifyWebhook(_ webhookId: String, options: [DiscordEndpoint.Options.WebhookOption],
-            callback: @escaping (DiscordWebhook?) -> () = {_ in }) {
-        DiscordEndpoint.modifyWebhook(webhookId, with: token, options: options, callback: callback)
-    }
-
-    /// Default implementation
-    public func removeGuildBan(for userId: String, on guildId: String, callback: ((Bool) -> ())? = nil) {
-        DiscordEndpoint.removeGuildBan(for: userId, on: guildId, with: token, callback: callback)
-    }
-
-    /// Default implementation.
-    public func removeGuildMemberRole(_ roleId: String, from userId: String, on guildId: String,
-                                      callback: ((Bool) -> ())?) {
-        DiscordEndpoint.removeGuildMemberRole(roleId, from: userId, on: guildId, with: token, callback: callback)
-    }
-
-    /// Default implementation
-    public func removeGuildRole(_ roleId: String, on guildId: String, callback: ((DiscordRole?) -> ())? = nil) {
-        DiscordEndpoint.removeGuildRole(roleId, on: guildId, with: token, callback: callback)
-    }
-
-    /// Default implementation.
-    public func sendMessage(_ message: DiscordMessage, to channelId: String,
-                            callback: ((DiscordMessage?) -> ())? = nil) {
-        DiscordEndpoint.sendMessage(message, with: token, to: channelId, callback: callback)
-    }
-
-    /// Default implementation
-    public func triggerTyping(on channelId: String, callback: ((Bool) -> ())? = nil) {
-        DiscordEndpoint.triggerTyping(on: channelId, with: token, callback: callback)
     }
 }
